moved audio context to file
