+++ src/components/Context.js
@@ -2,37 +2,9 @@
 /* eslint-disable react/state-in-constructor */
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { Synth, Envelope, Loop, Transport } from 'tone'
+import { toggleOnOff, setTransportBPM, setTransportTimeSig } from './AudioContext'
 
 const MetronomeContext = React.createContext()
-
-const synth = new Synth({
-  oscillator: {
-    type: 'triangle',
-  },
-  envelope: {
-    attack: 0.01,
-    decay: 0.05,
-    sustain: 0.0,
-    release: 1,
-  },
-}).toMaster()
-
-const env = new Envelope()
-
-env.set(0.0001, 5, 0.01, 0.001, 0.1, 0)
-
-const loop = new Loop((time) => {
-  Transport.setLoopPoints(0, '1m')
-  Transport.loop = true
-  const index = Transport.getTicksAtTime(time) / 192 + 1
-  console.log('Beat:', index)
-  if (index === 1) {
-    synth.triggerAttackRelease('C6', '16n')
-  } else {
-    synth.triggerAttackRelease('C5', '16n')
-  }
-}, '4n')
 
 export class Provider extends Component {
   state = {
@@ -48,7 +20,7 @@ export class Provider extends Component {
       }),
       () => {
         const { bpm } = this.state
-        Transport.bpm.value = bpm
+        setTransportBPM(bpm)
       }
     )
   }
@@ -60,7 +32,7 @@ export class Provider extends Component {
       }),
       () => {
         const { bpm } = this.state
-        Transport.bpm.value = bpm
+        setTransportBPM(bpm)
       }
     )
   }
@@ -72,7 +44,7 @@ export class Provider extends Component {
       }),
       () => {
         const { timeSignature } = this.state
-        Transport.timeSignature = timeSignature
+        setTransportTimeSig(timeSignature)
       }
     )
   }
@@ -80,12 +52,7 @@ export class Provider extends Component {
   handleStartLoop = (e) => {
     e.preventDefault()
     const { isPlaying } = this.state
-    if (!isPlaying) {
-      loop.start(0)
-      Transport.start()
-    } else {
-      Transport.stop()
-    }
+    toggleOnOff(isPlaying)
     this.setState((prevState) => ({
       isPlaying: !prevState.isPlaying,
     }))
